move varlisttoolbar to own script
add varlisttoolbar and sorted features to diseasevarlist
