+++ src/lib/components/var-list/VarList.js
@@ -1,29 +1,18 @@
 import React, { useEffect, useState } from "react";
 
-import {
-  faArrowDownAZ,
-  faArrowUpZA,
-  faArrowDown19,
-  faArrowUp91,
-  faXmark,
-} from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Add } from "@mui/icons-material";
-import { ToggleButton, ToggleButtonGroup } from "@mui/material";
 import _ from "lodash";
-import { ListGroup, Button, Alert, Navbar, InputGroup } from "react-bootstrap";
+import { ListGroup, Button, Alert } from "react-bootstrap";
 
 import { VarItem } from "./VarItem";
+import { VarListToolbar } from "./VarListToolbar";
 import { VarSet } from "./VarSet";
-import {
-  SELECTION_MODES,
-  VAR_SORT,
-  VAR_SORT_ORDER,
-} from "../../constants/constants";
+import { SELECTION_MODES, VAR_SORT } from "../../constants/constants";
 import { useDataset, useDatasetDispatch } from "../../context/DatasetContext";
 import { LoadingSpinner } from "../../utils/LoadingIndicators";
 import { useFetch } from "../../utils/requests";
 
+// @TODO: optimize
 const useVarMean = (varKeys, enabled = false) => {
   const ENDPOINT = "matrix/mean";
   const dataset = useDataset();
@@ -58,6 +47,7 @@ function DiseaseVarList({ makeListItem }) {
   const dataset = useDataset();
   const dispatch = useDatasetDispatch();
   const [diseaseVars, setDiseaseVars] = useState([]);
+  const [sortedDiseaseVars, setSortedDiseaseVars] = useState([]);
   const [params, setParams] = useState({
     url: dataset.url,
     col: dataset.varNamesCol,
@@ -71,24 +61,62 @@ function DiseaseVarList({ makeListItem }) {
     });
   }, [dataset.selectedDisease]);
 
-  const { fetchedData, isPending, serverError } = useFetch(ENDPOINT, params, {
+  const diseaseData = useFetch(ENDPOINT, params, {
     enabled: !!params.diseaseId,
   });
 
   useEffect(() => {
-    if (!isPending && !serverError) {
-      setDiseaseVars(fetchedData);
+    if (!diseaseData.isPending && !diseaseData.serverError) {
+      setDiseaseVars(diseaseData.fetchedData);
     }
-  }, [fetchedData, isPending, serverError]);
+  }, [diseaseData.fetchedData, diseaseData.isPending, diseaseData.serverError]);
 
-  const diseaseVarList = _.map(diseaseVars, (item) => {
+  const varMeans = useVarMean(
+    diseaseVars,
+    diseaseVars && dataset.varSort.disease.sort === VAR_SORT.MATRIX
+  );
+
+  useEffect(() => {
+    if (dataset.varSort.disease.sort === VAR_SORT.MATRIX) {
+      if (!varMeans.isPending && !varMeans.serverError) {
+        setSortedDiseaseVars(
+          _.orderBy(
+            diseaseVars,
+            (o) => {
+              return varMeans.fetchedData[o.name];
+            },
+            dataset.varSort.disease.sortOrder
+          )
+        );
+      }
+    } else if (dataset.varSort.disease.sort === VAR_SORT.NAME) {
+      setSortedDiseaseVars(
+        _.orderBy(diseaseVars, "name", dataset.varSort.disease.sortOrder)
+      );
+    } else {
+      setSortedDiseaseVars(diseaseVars);
+    }
+  }, [
+    dataset.varSort.disease.sort,
+    dataset.varSort.disease.sortOrder,
+    diseaseVars,
+    varMeans.fetchedData,
+    varMeans.isPending,
+    varMeans.serverError,
+  ]);
+
+  const diseaseVarList = _.map(sortedDiseaseVars, (item) => {
     return makeListItem(item, true);
   });
+
+  const isPending =
+    diseaseData.isPending ||
+    (varMeans.isPending && dataset.varSort.disease.sort === VAR_SORT.MATRIX);
 
   return (
     <>
       {dataset.selectedDisease &&
-        (!isPending && !fetchedData?.length ? (
+        (!isPending && !diseaseVars?.length ? (
           <>
             <div className="d-flex justify-content-between mt-3">
               <h5>Disease genes</h5>
@@ -111,109 +139,14 @@ function DiseaseVarList({ makeListItem }) {
               </Button>
             </div>
             <p>{dataset.selectedDisease?.name}</p>
-            <ListGroup>
-              {isPending ? <LoadingSpinner /> : diseaseVarList}
-            </ListGroup>
+            <VarListToolbar varType="disease" />
+            <div className="position-relative">
+              {isPending && <LoadingSpinner />}
+              <ListGroup>{diseaseVarList}</ListGroup>
+            </div>
           </>
         ))}
     </>
-  );
-}
-
-function VarListToolbar() {
-  const dataset = useDataset();
-  const dispatch = useDatasetDispatch();
-  const [sort, setSort] = useState(dataset.varSort.var.sort);
-  const [nameSortOrder, setNameSortOrder] = useState(
-    dataset.varSort.var.sortOrder
-  );
-  const [matrixSortOrder, setMatrixSortOrder] = useState(
-    dataset.varSort.var.sortOrder
-  );
-
-  const handleSort = (sortValue, sortOrder, setSortOrder) => {
-    if (sort !== sortValue) {
-      setSort(sortValue);
-      dispatch({
-        type: "set.varSort",
-        var: "var",
-        sort: sortValue,
-        sortOrder: sortOrder,
-      });
-    } else {
-      const newSortOrder =
-        sortOrder === VAR_SORT_ORDER.ASC
-          ? VAR_SORT_ORDER.DESC
-          : VAR_SORT_ORDER.ASC;
-      setSortOrder(newSortOrder);
-      dispatch({
-        type: "set.varSort",
-        var: "var",
-        sort: sortValue,
-        sortOrder: newSortOrder,
-      });
-    }
-  };
-
-  return (
-    <Navbar className="var-list-toolbar">
-      <InputGroup>
-        <InputGroup.Text>Sort by:</InputGroup.Text>
-
-        <ToggleButtonGroup
-          aria-label="Sort feature by"
-          size="small"
-          className="mh-100"
-        >
-          <ToggleButton
-            value={VAR_SORT.NAME}
-            aria-label="alphabetical"
-            title="Sort alphabetically"
-            selected={sort === VAR_SORT.NAME}
-            onChange={() => {
-              handleSort(VAR_SORT.NAME, nameSortOrder, setNameSortOrder);
-            }}
-          >
-            {nameSortOrder === VAR_SORT_ORDER.ASC ? (
-              <FontAwesomeIcon icon={faArrowDownAZ} />
-            ) : (
-              <FontAwesomeIcon icon={faArrowUpZA} />
-            )}
-          </ToggleButton>
-
-          <ToggleButton
-            value={VAR_SORT.MATRIX}
-            aria-label="matrix value"
-            title="Sort by matrix value"
-            selected={sort === VAR_SORT.MATRIX}
-            onChange={() => {
-              handleSort(VAR_SORT.MATRIX, matrixSortOrder, setMatrixSortOrder);
-            }}
-          >
-            {matrixSortOrder === VAR_SORT_ORDER.ASC ? (
-              <FontAwesomeIcon icon={faArrowDown19} />
-            ) : (
-              <FontAwesomeIcon icon={faArrowUp91} />
-            )}
-          </ToggleButton>
-          <ToggleButton
-            value="none"
-            aria-label="none"
-            title="No sorting"
-            onClick={() => {
-              setSort(VAR_SORT.NONE);
-              dispatch({
-                type: "set.varSort.sort",
-                var: "var",
-                sort: VAR_SORT.NONE,
-              });
-            }}
-          >
-            <FontAwesomeIcon icon={faXmark} />
-          </ToggleButton>
-        </ToggleButtonGroup>
-      </InputGroup>
-    </Navbar>
   );
 }
 
@@ -377,6 +310,9 @@ export function VarNamesList({
     return setName;
   };
 
+  const isPending =
+    varMeans.isPending && dataset.varSort.var.sort === VAR_SORT.MATRIX;
+
   return (
     <div className="position-relative">
       <div className="overflow-auto mt-3">
@@ -419,19 +355,22 @@ export function VarNamesList({
             </Button>
           </div>
         </div>
-        {!varList.length ? (
-          <Alert variant="light">Search for a feature.</Alert>
-        ) : (
-          <>
-            <VarListToolbar />
+        <>
+          {!varList.length ? (
+            <Alert variant="light">Search for a feature.</Alert>
+          ) : (
             <>
-              {dataset.varSort.var.sort !== VAR_SORT.NONE &&
-                varMeans.isPending && <LoadingSpinner />}
-              <ListGroup>{varList}</ListGroup>
+              <VarListToolbar />
+              <div className="position-relative">
+                {isPending && <LoadingSpinner />}
+                <ListGroup>{varList}</ListGroup>
+              </div>
             </>
-          </>
-        )}
-        {showDiseaseVarList && <DiseaseVarList makeListItem={makeListItem} />}
+          )}
+        </>
+        <>
+          {showDiseaseVarList && <DiseaseVarList makeListItem={makeListItem} />}
+        </>
       </div>
     </div>
   );
