+++ src/lib/utils/requests.js
@@ -1,5 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { useDebounce } from "@uidotdev/usehooks";
+import _ from "lodash";
 
 import { parseError } from "./errors";
 
@@ -49,9 +50,9 @@ export async function fetchData(
 export const useFetch = (endpoint, params, opts = {}, apiUrl = null) => {
   const { enabled = true } = opts;
   const {
-    data: fetchedData,
-    isLoading: isPending,
-    error: serverError,
+    data: fetchedData = null,
+    isLoading: isPending = false,
+    error: serverError = null,
   } = useQuery({
     queryKey: [endpoint, params],
     queryFn: ({ signal }) => {
@@ -78,9 +79,9 @@ export const useDebouncedFetch = (
   const debouncedParams = useDebounce(params, delay);
 
   const {
-    data: fetchedData,
-    isLoading: isPending,
-    error: serverError,
+    data: fetchedData = null,
+    isLoading: isPending = false,
+    error: serverError = null,
   } = useQuery({
     queryKey: [endpoint, debouncedParams],
     queryFn: ({ signal }) => {
